requirements.txt
https://github.com/damianmoore/matomo-monorail/releases/download/v0.1.2/matomo_monorail-0.1.2.tar.gz
Django==3.0.4
django-filer==1.7.0
django-filter==2.2.0
django-pagedown==2.0.3
easy-thumbnails==2.7
gunicorn==19.9.0
psycopg2-binary==2.8.4
requests==2.20.0
